# services/api-gateway/docs/swagger.yaml
basePath: /v1
consumes:
- application/json
definitions:
  httpx.ErrorBody:
    properties:
      code:
        example: VALIDATION_ERROR
        type: string
      message:
        example: invalid request
        type: string
    type: object
  httpx.ErrorResponse:
    properties:
      error:
        $ref: '#/definitions/httpx.ErrorBody'
    type: object
  main.adminAccountListResponse:
    properties:
      admin_accounts:
        items:
          $ref: '#/definitions/main.adminAccountView'
        type: array
    type: object
  main.adminAccountView:
    properties:
      created_at:
        type: string
      display_name:
        type: string
      id:
        type: string
      is_self:
        type: boolean
      username:
        type: string
    type: object
...
  main.adminPhoneListResponse:
    properties:
      admin_phones:
        items:
          $ref: '#/definitions/main.adminPhoneView'
        type: array
    type: object
  main.adminPhoneView:
    properties:
      created_at:
        type: string
      id:
        type: string
      is_self:
        type: boolean
      label:
        type: string
      phone_masked:
        type: string
    type: object
...
  main.createAdminAccountBody:
    properties:
      display_name:
        type: string
      password:
        type: string
      username:
        type: string
    type: object
  main.createAdminPhoneBody:
    properties:
      label:
        type: string
      phone:
        type: string
    type: object
...
  main.deleteAdminPhoneResponse:
    properties:
      id:
        type: string
      ok:
        type: boolean
    type: object
...
  main.updateAdminAccountBody:
    properties:
      current_password:
        type: string
      display_name:
        type: string
      new_password:
        type: string
      username:
        type: string
    type: object
info:
  contact: {}
  description: Public API gateway for Gas Tam De customers and administrators.
  title: Gas Tam De API
  version: "1.0"
paths:
  /admin/admin-accounts:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.adminAccountListResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List administrator accounts
      tags:
      - Admin - Authentication
    post:
      consumes:
      - application/json
      parameters:
      - description: Administrator account
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/main.createAdminAccountBody'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/main.adminAccountView'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create an administrator account
      tags:
      - Admin - Authentication
  /admin/admin-accounts/{id}:
    patch:
      consumes:
      - application/json
      parameters:
      - description: Administrator account ID
        in: path
        name: id
        required: true
        type: string
      - description: Account changes
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/main.updateAdminAccountBody'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.adminAccountView'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update an administrator account
      tags:
      - Admin - Authentication
  /admin/admin-phones:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.adminPhoneListResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List administrator phone numbers
      tags:
      - Admin - Authentication
    post:
      consumes:
      - application/json
      parameters:
      - description: Phone and optional label
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/main.createAdminPhoneBody'
      produces:
      - application/json
      responses:
        "200":
          description: Existing entry
          schema:
            $ref: '#/definitions/main.adminPhoneView'
        "201":
          description: Created
          schema:
            $ref: '#/definitions/main.adminPhoneView'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Add an administrator phone number
      tags:
      - Admin - Authentication
  /admin/admin-phones/{id}:
    delete:
      parameters:
      - description: Administrator phone ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.deleteAdminPhoneResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Remove an administrator phone number
      tags:
      - Admin - Authentication
...
produces:
- application/json
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: 'Enter the access token as: Bearer <token>'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"